requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions==1.12.0
protobuf==3.12.2
onnxruntime==1.13.1
transformers==4.24.0
scipy==1.8.0
